# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TG Restaurant ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /finance/expenses:
    get:
      operationId: listExpenses
      tags: [finance]
      summary: List expenses
      parameters:
        - name: branchId
          in: query
          schema:
            type: integer
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [finance]
      summary: Log expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /finance/expenses/{id}:
    delete:
      operationId: deleteExpense
      tags: [finance]
      summary: Delete expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /finance/summary:
    get:
      operationId: getFinanceSummary
      tags: [finance]
      summary: Get financial summary (income, expenses, profit)
      parameters:
        - name: branchId
          in: query
          schema:
            type: integer
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Financial summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinanceSummary"
  /finance/revenue-trend:
    get:
      operationId: getRevenueTrend
      tags: [finance]
      summary: Daily revenue trend for chart
      parameters:
        - name: branchId
          in: query
          schema:
            type: integer
        - name: days
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Revenue trend
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RevenueTrendPoint"

  # ─── INVENTORY ────────────────────────────────────────────
...
components:
  schemas:
...
    Expense:
      type: object
      required: [id, branchId, category, amountAed, description, createdAt]
      properties:
        id:
          type: integer
        branchId:
          type: integer
        category:
          type: string
        amountAed:
          type: number
        description:
          type: string
        loggedBy:
          type: ["integer", "null"]
        loggedByName:
          type: ["string", "null"]
        createdAt:
          type: string
    ExpenseInput:
      type: object
      required: [branchId, category, amountAed, description]
      properties:
        branchId:
          type: integer
        category:
          type: string
        amountAed:
          type: number
        description:
          type: string
    FinanceSummary:
      type: object
      required: [totalRevenue, totalExpenses, netProfit, orderCount]
      properties:
        totalRevenue:
          type: number
        totalExpenses:
          type: number
        netProfit:
          type: number
        orderCount:
          type: integer
        avgOrderValue:
          type: number
        branchId:
          type: ["integer", "null"]
        from:
          type: ["string", "null"]
        to:
          type: ["string", "null"]
    RevenueTrendPoint:
      type: object
      required: [date, revenue, orderCount]
      properties:
        date:
          type: string
        revenue:
          type: number
        orderCount:
          type: integer

    # INVENTORY
...