# lib/api-zod/src/generated/types/revenueTrendPoint.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * TG Restaurant ERP API
 * OpenAPI spec version: 0.1.0
 */

export interface RevenueTrendPoint {
  date: string;
  revenue: number;
  orderCount: number;
}